+++ src/main/java/su26sd09/su26sd09/controller/NhanVienDatPhongController.java
@@ -825,7 +825,7 @@ public class NhanVienDatPhongController {
         }
         if("Chuyen Khoan".equalsIgnoreCase(phuongthuc)){
             String baseUrl = request.getScheme() + "://"+request.getServerName() + ":"+request.getServerPort();
-            String vnPayUrl = vnpayService.createOrder(soTien.intValue(),id,"ThuThemDichVu",baseUrl);
+            String vnPayUrl = vnpayService.createOrder(soTien.longValue(),id,"ThuThemDichVu",baseUrl);
             return "redirect:"+vnPayUrl;
         }
         ThanhToan tt = new ThanhToan();
