+++ src/main/java/su26sd09/su26sd09/controller/NhanVienHoanTienController.java
@@ -107,7 +107,7 @@ public class NhanVienHoanTienController {
 
         String baseUrl = request.getScheme() + "://" + request.getServerName() + ":" + request.getServerPort();
         String vnpayUrl = vnpayService.createRefundOrder(
-                hd.getId(), hd.getSoTienHoan().intValue(), baseUrl);
+                hd.getId(), hd.getSoTienHoan().longValue(), baseUrl);
         return "redirect:" + vnpayUrl;
     }
